+++ altapay/transaction.py
@@ -26,4 +26,29 @@ class Transaction(Resource):
 
         return cls(
             response['@version'], response['Header'],
-            response['Body']['Transactions']['Transaction'])
+            response['Body']['Transactions']['Transaction'], api=api)
+
+    def capture(self, **kwargs):
+        """
+        Capture a reservation on a transaction.
+
+        :param \*\*kwargs: used for optional capture parameters, see the
+            AltaPay documentation for a full list.
+            Note that you will need to use lists and dictionaries to map the
+            URL structures from the AltaPay documentation into these kwargs.
+        """
+        parameters = {
+            'transaction_id': self.transaction_id
+        }
+
+        parameters.update(kwargs)
+
+        response = self.api.get(
+            'API/captureReservation', parameters=parameters)['APIResponse']
+
+        super(Transaction, self).__setattr__(
+            '__data__', response['Body']['Transactions']['Transaction'])
+
+        return Transaction(
+            response['@version'], response['Header'],
+            response['Body']['Transactions']['Transaction'], api=self.api)