+++ tests/test_transaction.py
@@ -34,3 +34,23 @@ class PaymentTest(TestCase):
 
         with self.assertRaises(MultipleResourcesError):
             Transaction.find('TEST-TRANSACTION-ID', self.api)
+
+    @responses.activate
+    def test_capture_transaction_simple(self):
+        responses.add(
+            responses.GET, self.get_api_url('API/payments'),
+            body=self.load_xml_response('200_find_transaction_single.xml'),
+            status=200, content_type='application/xml')
+
+        transaction = Transaction.find('TEST-TRANSACTION-ID', self.api)
+
+        self.assertEqual(transaction.captured_amount, 0.0)
+
+        responses.add(
+            responses.GET, self.get_api_url('API/captureReservation'),
+            body=self.load_xml_response('200_capture_response.xml'),
+            status=200, content_type='application/xml')
+
+        transaction = transaction.capture()
+
+        self.assertEqual(transaction.captured_amount, 13.29)